+++ ChangeLog.md
@@ -1,3 +1,8 @@
 # Changelog for artificial-haskell-test
 
-## Unreleased changes
+## 0.0.0.2
+  - Fix mempry leak and CPU choke
+  - Add direct village coordinate search
+  - Cmd-line options for specific processing type
+  - Introduce few datatypes
+  - Concurrent scroll parsing